Split static assets into separate files
