+++ packages/doenetml-worker-javascript/src/test/chemistry/orbitalDiagram.test.ts
@@ -1,10 +1,5 @@
 import { describe, expect, it, vi } from "vitest";
 import { createTestCore } from "../utils/test-core";
-import { updateMathInputValue } from "../utils/actions";
-import { atomDatabase } from "@doenet/static-assets";
-//@ts-ignore
-import me from "math-expressions";
-import { superSubscriptsToUnicode } from "../../utils/math";
 
 const Mock = vi.fn();
 vi.stubGlobal("postMessage", Mock);
