+++ packages/lsp-tools/test/doenet-auto-schema-check.test.ts
@@ -1,9 +1,6 @@
 import { describe, expect, it } from "vitest";
 import util from "util";
 
-import { filterPositionInfo, DastMacro, DastElement } from "@doenet/parser";
-import { DoenetSourceObject, isOldMacro } from "../src/doenet-source-object";
-import { doenetSchema } from "@doenet/static-assets";
 import { AutoCompleter } from "../src";
 
 const origLog = console.log;
