+++ packages/static-assets/src/atom-database.ts
@@ -1,0 +1,3 @@
+//@ts-ignore
+import _atomDatabase from "./data/atom-database.csv";
+export const atomDatabase = _atomDatabase as Record<string, string>[];
