+++ packages/static-assets/src/entity-map.ts
@@ -1,0 +1,2 @@
+import _ENTITY_MAP from "./generated/entity-map.json";
+export const ENTITY_MAP: Record<string, string> = _ENTITY_MAP;
