+++ packages/static-assets/src/math-assets.ts
@@ -1,0 +1,2 @@
+import mathAssets from "./generated/math-assets.json";
+export { mathAssets };
